Float64 everywhere (#65)
* Added a method to convert mgl64.Mat4 to 32bits

* Added error in case 64bits data is passed

* Converted everything to float64

+++ pkg/input/game_controller.go
@@ -57,7 +57,7 @@ type GameController interface {
 	ButtonPressed(button ControllerButton) bool
 	ButtonReleased(button ControllerButton) bool
 	ButtonDown(button ControllerButton) bool
-	AxisValue(axis ControllerAxis) float32
+	AxisValue(axis ControllerAxis) float64
 	AxisDigitalValue(axis ControllerAxis) int
 	SetMapping(mapping *GameControllerMapping)
 	Description() string
